feat: download from OpenAPI document URL (if available)
