+++ packages/api-reference/src/features/DownloadLink/index.ts
@@ -1,0 +1,1 @@
+export { default as DownloadLink } from './DownloadLink.vue'
