+++ packages/api-reference/src/helpers/provideSymbols.ts
@@ -1,13 +1,20 @@
-import type { OpenAPIV3 } from '@scalar/openapi-parser'
+import type { OpenAPIV3, OpenAPIV3_1 } from '@scalar/openapi-parser'
 import type { InjectionKey } from 'vue'
 
 /**
  * Collection of symbols to use with provide inject
  */
 export const GLOBAL_SECURITY_SYMBOL = Symbol() as InjectionKey<
-  () => OpenAPIV3.SecurityRequirementObject[] | undefined
+  () =>
+    | OpenAPIV3.SecurityRequirementObject[]
+    | OpenAPIV3_1.SecurityRequirementObject
+    | undefined
 >
 
 export const HIDE_DOWNLOAD_BUTTON_SYMBOL = Symbol() as InjectionKey<
   () => boolean | undefined
+>
+
+export const OPENAPI_DOCUMENT_URL_SYMBOL = Symbol() as InjectionKey<
+  () => string | undefined
 >
